Zmiana sposobu logowania , poprawki
git-svn-id: https://localhost:8443/svn/hm-mde-1@480 fca3399c-2208-499e-aaa1-2cf7fb07c25e

+++ trunk/edr/src/corelib/ChartPointer.cpp
@@ -93,7 +93,7 @@ std::string formatNumber( float number )
 }
 void ChartPointer::update(double targetTime){
 	
-	int frame=targetTime*data->getFPS();
+	int frame = static_cast<int>(targetTime * data->getFPS());
 	//if(frame<data->getRNumber()-1){
 	trans->setMatrix(osg::Matrix::translate(data->getNormalizedXValue(frame) * (endPoint.x()-startPoint.x()),0.0f,0.0f));
 	if(labelVisable){
